runs/2026-01-16__ge-market-insights/ge-pipeline/requirements.txt
# Core
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-pptx>=0.6.23
Pillow>=9.2
PyYAML>=6.0
typer>=0.12.0
Jinja2>=3.1.0

# Google BigQuery
google-cloud-bigquery>=3.11.0
# Optional but recommended for faster loads -> pandas via pyarrow
pyarrow>=12.0.0
# Ensure correct dtype handling for BigQuery
db-dtypes>=1.1.0
# Optional storage API for higher throughput
google-cloud-bigquery-storage>=2.22.0

# Jupyter
jupyter>=1.0.0

# Testing
pytest>=8.0.0
